Add parseInt as per specifications
Source's parseInt requires that both first and second positional
arguments, the inputString and radix be provided. This is in contrast to
the native implementation of parseInt where the second radix argument
can be omitted. As a result, a wrapper function `parse_int` in `misc`
was defined to raise errors for bad arguments.